requirements.txt
jupyterlab
python-dotenv
tqdm
beautifulsoup4
playwright
openai
gradio
flask
fpdf2
gunicorn
markdown
pdfkit